+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.7.0]
+### Changes
+- Removed duplicate notifications during NPM auth flow
+- Changed no github auth message to warning
+
+### Fixed
+- Orion packages no longer show if the user is not in the organization (this may have already worked)
+
+## [0.5.0], [0.6.0]
+### Changes
+- Fixes various configuration files
+
 ## [0.4.0]
 ### Changes
 - Properly setup esbuild to ensure that smallest possible extension is compiled